(#40): Swiper Slide 컴포넌트 분리
